# back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: "Ugram API"
  description: "A Ugram API documentation by team 12"
  version: "1.0"
host: "localhost:5000"
basePath: "/"
consumes:
- "application/json"
produces:
- "application/json"
...
schemes:
- "http"
paths:
...
  /post/id/{id_post}:
    get:
      tags:
      - "post"
      summary: "Get a post infos"
      parameters:
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "The post infos"
          schema:
            $ref: "#/definitions/ResponseObjectPost"
  /post/person/{id_person}/{offset}:
    get:
      tags:
      - "post"
      summary: "Get all posts related to a user"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
  /post/hashtag/{id_hashtag}:
    get:
      tags:
      - "post"
      summary: "Get posts with this hashtag"
      parameters:
      - in: "path"
        name: "id_hashtag"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
  /post/date/{offset}:
    get:
      tags:
      - "post"
      summary: "Get most recent posts"
      parameters:
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
  /post/subnews/{offset}:
    get:
      tags:
      - "post"
      summary: "Get my subscriptions most recent posts"
      parameters:
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
  /post/random/{offset}:
    get:
      tags:
      - "post"
      summary: "Get random posts"
      parameters:
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
  /post/post_image/{id_post}:
    get:
      tags:
      - "post"
      summary: "Get post image"
      produces:
      - "image/png"
      parameters:
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Image file"
          schema:
            type: "file"
  /post/post_image/{id_person}/{id_post}:
    post:
      tags:
      - "post"
      summary: "Upload post image"
      consumes:
       - "multipart/form-data"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      - in: "formData"
        name: "image"
        type: "file"
      responses:
        "200":
          description: "OK"
  /post:
    post:
      tags:
      - "post"
      summary: "Create a new post"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos"
        required: true
        schema:
          $ref: "#/definitions/PostInputCreate"
      responses:
        "201":
          description: "Id of the newly created post"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
    put:
      tags:
      - "post"
      summary: "Update a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos to update"
        required: true
        schema:
          $ref: "#/definitions/PostInputUpdate"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "post"
      summary: "Delete a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the post to delete"
        required: true
        schema:
          $ref: "#/definitions/PostInputDelete"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
...
definitions:
...
  PersonName:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      username_person:
        type: "string"
        default: "thejohndoe_"
...
  PostInputCreate:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
  PostInputUpdate:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
  PostInputDelete:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
        description: "Id of the logged in user"
  PostOutput:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      person:
        $ref: "#/definitions/PersonName"
      isliked_post:
        type: "boolean"
        default: "true"
      likes_post:
        type: "integer"
        default: "3"
      comments_post:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
...
  ResponseObjectEmpty:
    type: "object"
    properties:
      response:
        type: "object"
        default: null
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectInt:
    type: "object"
    properties:
      response:
        type: "integer"
        default: 1
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectPost:
    type: "object"
    properties:
      response:
        $ref: "#/definitions/PostOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectPostArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/PostOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"  
...